Move sandbox normalization helpers to private module
Keep shared on-demand sandbox normalization logic out of the public client and declaration modules by centralizing it in a private normalization module.
